Activity vs extensions refactoring
