+++ src/libraries/Core/Microsoft.Agents.Core/Models/IActivityExtensions.cs
@@ -1,8 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
-using System.Collections.Generic;
 using System.Text.Json;
 using Microsoft.Agents.Core.Serialization;
 
@@ -21,40 +19,6 @@ namespace Microsoft.Agents.Core.Models
         public static string ToJson(this IActivity activity)
         {
             return JsonSerializer.Serialize(activity, ProtocolJsonSerializer.SerializationOptions);
-        }
-
-        /// <summary>
-        /// Resolves the mentions from the entities of this activity.
-        /// </summary>
-        /// <returns>The array of mentions; or an empty array, if none are found.</returns>
-        /// <remarks>This method is defined on the <see cref="Microsoft.Agents.Core.Models.Activity"/> class, but is only intended
-        /// for use with a message activity, where the activity <see cref="Microsoft.Agents.Core.Models.Activity.Type"/> is set to
-        /// <see cref="Microsoft.Agents.Core.Models.ActivityTypes.Message"/>.</remarks>
-        /// <seealso cref="Microsoft.Agents.Core.Models.Mention"/>
-        public static Mention[] GetMentions(this IActivity activity)
-        {
-            var result = new List<Mention>();
-            if (activity.Entities != null)
-            {
-                foreach (var entity in activity.Entities)
-                {
-                    if (entity is Mention mention)
-                    {
-                        result.Add(mention);
-                    }
-                }
-            }
-            return [.. result];
-        }
-
-        /// <summary>
-        /// Remove recipient mention text from Text property.
-        /// Use with caution because this function is altering the text on the Activity.
-        /// </summary>
-        /// <returns>new .Text property value.</returns>
-        public static string RemoveRecipientMention<T>(this T activity) where T : IActivity
-        {
-            return activity.RemoveMentionText(activity.Recipient?.Id);
         }
 
         /// <summary>
@@ -117,63 +81,6 @@ namespace Microsoft.Agents.Core.Models
             {
                 return false;
             }
-        }
-
-        /// <summary>
-        /// Determines whether the specified activity represents a targeted activity treatment.
-        /// </summary>
-        /// <remarks>Use this method to identify activities that are specifically marked as targeted
-        /// treatments. This can be useful for filtering or processing activities based on their treatment
-        /// type.</remarks>
-        /// <param name="activity">The activity to evaluate for targeted treatment. Cannot be null.</param>
-        /// <returns>true if the activity contains an entity of type ActivityTreatment with a treatment of Targeted; otherwise,
-        /// false.</returns>
-        public static bool IsTargetedActivity(this IActivity activity)
-        {
-            if (activity.Entities == null || activity.Entities.Count == 0)
-            {
-                return false;
-            }
-
-            foreach (var entity in activity.Entities)
-            {
-                if (entity.Type == EntityTypes.ActivityTreatment && entity is ActivityTreatment treatment && treatment.Treatment == ActivityTreatmentTypes.Targeted)
-                {
-                    return true;
-                }
-            }
-
-            return false;
-        }
-
-        /// <summary>
-        /// Marks the specified activity as targeted by adding a targeted treatment entity to its collection.
-        /// </summary>
-        /// <remarks>This method ensures that the activity's Entities collection is initialized before
-        /// adding the targeted treatment. Use this extension to indicate that an activity should be processed or
-        /// interpreted as targeted in downstream workflows.</remarks>
-        /// <param name="activity">The activity to be marked as targeted. Cannot be null.</param>
-        /// <param name="user">The user to be set as the recipient.  This replaces Activity.Recipient if provided.</param>
-        public static IActivity MakeTargetedActivity(this IActivity activity, ChannelAccount user = null)
-        {
-            AssertionHelpers.ThrowIfNull(activity, nameof(activity));
-
-            if (activity.IsTargetedActivity())
-            {
-                return activity;
-            }
-
-            if (activity.Recipient == null && user == null)
-            {
-                throw new InvalidOperationException("Cannot mark activity as targeted because both the Activity.Recipient and `user` argument are null. At least one must be provided.");
-            }
-
-            activity.Entities ??= [];
-            activity.Entities.Add(new ActivityTreatment() { Treatment = ActivityTreatmentTypes.Targeted });
-
-            activity.Recipient = user ?? activity.Recipient;
-
-            return activity;
         }
     }
 }
